# Refactoring of Timestamp