--- a/.sbtopts
+++ b/.sbtopts
@@ -2,5 +2,5 @@
 -J-XX:MaxInlineLevel=20
 -J-Xss2m
 -J-Xms512M
--J-Xmx4G
+-J-Xmx6G
 -J-XX:ReservedCodeCacheSize=256M