--- a/modules/aws/src/smithy4s/json/AwsSchematicJCodec.scala
+++ b/modules/aws/src/smithy4s/json/AwsSchematicJCodec.scala
@@ -42,17 +42,13 @@ private[aws] object AwsSchematicJCodec
   private val timestamp: JCodec[Timestamp] = new JCodec[Timestamp] {
     val expecting: String = "instant (epoch second)"
 
-    def decodeValue(cursor: Cursor, in: JsonReader): Timestamp =
-      Timestamp.fromEpochSecond(in.readDouble().toLong)
+    def decodeValue(cursor: Cursor, in: JsonReader): Timestamp = Timestamp(in.readDouble().toLong, 0)
 
-    def encodeValue(x: Timestamp, out: JsonWriter): Unit =
-      out.writeVal(x.epochSecond)
+    def encodeValue(x: Timestamp, out: JsonWriter): Unit = out.writeVal(x.epochSecond)
 
-    def decodeKey(in: JsonReader): Timestamp =
-      Timestamp.fromEpochSecond(in.readKeyAsDouble().toLong)
+    def decodeKey(in: JsonReader): Timestamp = Timestamp(in.readKeyAsDouble().toLong, 0)
 
-    def encodeKey(x: Timestamp, out: JsonWriter): Unit =
-      out.writeKey(x.epochSecond)
+    def encodeKey(x: Timestamp, out: JsonWriter): Unit = out.writeKey(x.epochSecond)
   }
 
 }