Goals for refactoring of `smithy4s.Timestamp`:
- Fix construction/parsing/serialization/formatting to be consistent between platforms and follow [the protocol standards](https://awslabs.github.io/smithy/1.0/spec/core/stream-traits.html#event-message-serialization)
- Minimize the memory footprint for all platforms
- More efficient parsing/formatting from/to string for all supported formats
- More efficient parsing/serialization from/to JSON for all supported formats

The range of allowed values reduced to be from `0000-01-01T00:00:00Z` to `9999-12-31T23:59:59Z`.

Below are results of benchmarks for parsing and serialization of array of 128 `java.time.Instant` values using JSON codec for `smithy4s.Timestamp` on Intel® Core™ i9-11900H with JDK 17.

Before:
```
[info] Benchmark                                                             (size)   Mode  Cnt       Score       Error   Units
[info] ArrayOfInstantsReading.smithy4sJson                                      128  thrpt    5   11671.896 ±   147.913   ops/s
[info] ArrayOfInstantsReading.smithy4sJson:·gc.alloc.rate                       128  thrpt    5    2071.824 ±    26.405  MB/sec
[info] ArrayOfInstantsReading.smithy4sJson:·gc.alloc.rate.norm                  128  thrpt    5  279281.183 ±     0.073    B/op
[info] ArrayOfInstantsReading.smithy4sJson:·gc.churn.PS_Eden_Space              128  thrpt    5    2046.870 ±     0.376  MB/sec
[info] ArrayOfInstantsReading.smithy4sJson:·gc.churn.PS_Eden_Space.norm         128  thrpt    5  275919.803 ±  3493.766    B/op
[info] ArrayOfInstantsReading.smithy4sJson:·gc.churn.PS_Survivor_Space          128  thrpt    5       0.100 ±     0.274  MB/sec
[info] ArrayOfInstantsReading.smithy4sJson:·gc.churn.PS_Survivor_Space.norm     128  thrpt    5      13.484 ±    37.148    B/op
[info] ArrayOfInstantsReading.smithy4sJson:·gc.count                            128  thrpt    5      20.000              counts
[info] ArrayOfInstantsReading.smithy4sJson:·gc.time                             128  thrpt    5      10.000                  ms
[info] ArrayOfInstantsWriting.smithy4sJson                                      128  thrpt    5   37616.648 ±   939.227   ops/s
[info] ArrayOfInstantsWriting.smithy4sJson:·gc.alloc.rate                       128  thrpt    5    2311.128 ±    57.287  MB/sec
[info] ArrayOfInstantsWriting.smithy4sJson:·gc.alloc.rate.norm                  128  thrpt    5   96667.132 ±     1.520    B/op
[info] ArrayOfInstantsWriting.smithy4sJson:·gc.churn.PS_Eden_Space              128  thrpt    5    2251.551 ±  1079.548  MB/sec
[info] ArrayOfInstantsWriting.smithy4sJson:·gc.churn.PS_Eden_Space.norm         128  thrpt    5   94203.531 ± 46229.876    B/op
[info] ArrayOfInstantsWriting.smithy4sJson:·gc.churn.PS_Survivor_Space          128  thrpt    5       0.071 ±     0.166  MB/sec
[info] ArrayOfInstantsWriting.smithy4sJson:·gc.churn.PS_Survivor_Space.norm     128  thrpt    5       2.954 ±     6.901    B/op
[info] ArrayOfInstantsWriting.smithy4sJson:·gc.count                            128  thrpt    5      22.000              counts
[info] ArrayOfInstantsWriting.smithy4sJson:·gc.time                             128  thrpt    5      12.000                  ms
```

After:
```
[info] Benchmark                                                             (size)   Mode  Cnt       Score       Error   Units
[info] ArrayOfInstantsReading.smithy4sJson                                      128  thrpt    5  110357.823 ±  2238.914   ops/s
[info] ArrayOfInstantsReading.smithy4sJson:·gc.alloc.rate                       128  thrpt    5    1908.421 ±    38.816  MB/sec
[info] ArrayOfInstantsReading.smithy4sJson:·gc.alloc.rate.norm                  128  thrpt    5   27208.923 ±     0.419    B/op
[info] ArrayOfInstantsReading.smithy4sJson:·gc.churn.PS_Eden_Space              128  thrpt    5    1944.579 ±   881.102  MB/sec
[info] ArrayOfInstantsReading.smithy4sJson:·gc.churn.PS_Eden_Space.norm         128  thrpt    5   27726.767 ± 12634.748    B/op
[info] ArrayOfInstantsReading.smithy4sJson:·gc.churn.PS_Survivor_Space          128  thrpt    5       0.075 ±     0.193  MB/sec
[info] ArrayOfInstantsReading.smithy4sJson:·gc.churn.PS_Survivor_Space.norm     128  thrpt    5       1.068 ±     2.752    B/op
[info] ArrayOfInstantsReading.smithy4sJson:·gc.count                            128  thrpt    5      19.000              counts
[info] ArrayOfInstantsReading.smithy4sJson:·gc.time                             128  thrpt    5       8.000                  ms
[info] ArrayOfInstantsWriting.smithy4sJson                                      128  thrpt    5  145260.769 ±  5517.860   ops/s
[info] ArrayOfInstantsWriting.smithy4sJson:·gc.alloc.rate                       128  thrpt    5    2322.867 ±    88.449  MB/sec
[info] ArrayOfInstantsWriting.smithy4sJson:·gc.alloc.rate.norm                  128  thrpt    5   25160.847 ±     0.378    B/op
[info] ArrayOfInstantsWriting.smithy4sJson:·gc.churn.PS_Eden_Space              128  thrpt    5    2353.869 ±  1079.483  MB/sec
[info] ArrayOfInstantsWriting.smithy4sJson:·gc.churn.PS_Eden_Space.norm         128  thrpt    5   25490.408 ± 11403.692    B/op
[info] ArrayOfInstantsWriting.smithy4sJson:·gc.churn.PS_Survivor_Space          128  thrpt    5       0.121 ±     0.191  MB/sec
[info] ArrayOfInstantsWriting.smithy4sJson:·gc.churn.PS_Survivor_Space.norm     128  thrpt    5       1.307 ±     2.058    B/op
[info] ArrayOfInstantsWriting.smithy4sJson:·gc.count                            128  thrpt    5      23.000              counts
[info] ArrayOfInstantsWriting.smithy4sJson:·gc.time                             128  thrpt    5      12.000                  ms
```