@kubukoz 2022-07-12
> The range of allowed values reduced to be from 0000-01-01T00:00:00Z to 9999-12-31T23:59:59Z.

@plokhotnyuk can you share more context on this? Looks like the default `Abitrary[Instant]` from scalacheck is now generating values that fall outside this range, so any user code relying on these arbitraries will break