@plokhotnyuk 2022-07-12
> @plokhotnyuk can you share more context on this? Looks like the default `Abitrary[Instant]` from scalacheck is now generating values that fall outside this range, so any user code relying on these arbitraries will break

The reason of decreasing the range of allowed values for `smithy4s.Timestamp` is preventing of loosing information during serialization to [standard formats](https://awslabs.github.io/smithy/1.0/spec/core/protocol-traits.html?highlight=timestamp%20format#timestampformat-trait).